// plio-api/plio/apis/EventController.ts
/* eslint-disable */

/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    Event,
    EventFromJSON,
    EventToJSON,
} from '../models/Event';
import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & Event;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & Event;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

/**
 * no description
 */
export const EventController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/plio/api/Event/{id}')
    }),
    /**
     */
    create: createEndpoint<Event, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/plio/api/Event')
    }),
    /**
     */
    read: createEndpoint<Event, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Event/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<Event>, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Event')
    }),
    /**
     */
    sparseUpdate: createEndpoint<Event, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/plio/api/Event/{id}')
    }),
    /**
     */
    update: createEndpoint<Event, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/plio/api/Event')
    }),
}